Potions, Pickups, and Pop Up Text
This adds a lot of changes, like pickups that drop from enemies, pop up text for damage numbers and pickups, and two of the potions. Use potions with F, use Q and E to switch potions, the number next to the health bar shows which potion is active. (0 for health, 1 for jump, 2 for damage) currently the damage potion isn't working, but the other two are